Dev/gerenciador_de_ofertas/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.3